Adds counter in the bubbles example

+++ bubbles/bubbles-backend/src/main/java/org/acme/Bubble.java
@@ -2,7 +2,7 @@ package org.acme;
 
 public class Bubble {
     
-    public Bubble(String color, String hostname, int requests) {
+    public Bubble(String color, String hostname, long requests) {
         this.color = color;
         this.hostname = hostname;
         this.requests = requests;
@@ -10,5 +10,5 @@ public class Bubble {
     
     public String color;
     public String hostname;
-    public int requests;
+    public long requests;
 }
